BOT-1009-core-additional-asserters-for-t - matchers moved to MatchFunctions.js
